requirements.txt
requests
PyQt5
PyQtWebEngine
PySocks
